lensmu/backend/requirements.txt
# =============================================================================
# VisionTranslate (lensmu) Backend Dependencies
# =============================================================================
#
# INSTALL IN TWO STEPS:
#
#   Step 1 — Core server (always works):
#     pip install -r requirements.txt
#
#   Step 2 — OCR engines (install whichever you want):
#     pip install -r requirements-ocr.txt
#
#   See requirements-ocr.txt for platform-specific PaddlePaddle instructions.
#
# The server starts even if no OCR engines are installed. Endpoints for
# missing engines will return a helpful error message instead of crashing.
# =============================================================================

# --- Web framework and server ---
fastapi>=0.104.0          # async web framework for building the API
uvicorn[standard]>=0.24.0 # ASGI server to run FastAPI (the [standard] extra
                          # includes uvloop and httptools for better performance)

# --- Image processing ---
Pillow>=10.0.0            # PIL fork for image decoding, cropping, and conversion
numpy>=1.24.0             # array operations used internally by both OCR engines

# --- Request parsing ---
python-multipart>=0.0.6   # required by FastAPI for form/file upload parsing
